# Replace channel-based DbHandler with high-performance Mutex implementation