--- a/.gitignore
+++ b/.gitignore
@@ -14,3 +14,5 @@ __pycache__/
 
 # Benchmarks
 benchmarks/.venv/
+
+sqlite.db*