--- a/src/session/mod.rs
+++ b/src/session/mod.rs
@@ -5,4 +5,4 @@ pub mod db_handler;
 
 pub use state::{SessionState, PreparedStatement, Portal, GLOBAL_QUERY_CACHE};
 pub use pool::{SqlitePool, PooledConnection};
-pub use db_handler::DbHandler;
+pub use db_handler::{DbHandler, DbResponse};